[Java] Extend ATJ to generate MV classes.
Now ATJ generates, for each type of MV results returned by ACL2 functions that
are translated to Java, a Java class that represents tuples of such MV results.
This is another step towards supporting more directly and efficiently MV values
in the generated Java code. These newly generated classes are not used by the
rest of the Java code yet; they will be used as more support is added.