[Java] Fix bug in ATJ.
Besides MV classes for the result types of multi-valued functions, in general it
is also necessary to generate MV classes for multiple values that may be created
on the fly via MV calls. These cases actually do not occur in the current tests,
but they may occur in general and thus they should be handled.