Made minor tweaks to define and translate.
The tweak to define is to change extend-define-guts-alist to replace a
call of cons on two quoted objects with a quoted cons.  Initial
timings seemed to suggest a slight speedup from this change, but later
ones showed little if any speedup.  Still, it seems like a reasonable
change, as it avoids a call to translate11 on the cons call, so I'll
keep it.

The change to translate is actually to the function translate11, and
is primarily to simplify the code a bit.  It combines two very similar
COND clauses into one, and it avoids the need to keep in sync with the
function, primitive-event-macros.  Also, it uses hons-get in an
attempt to improve efficiency, though the improvement seems to be
negligible.