[Java] Extend ATJ user documentation.
Discuss the possiblity of stack overflow errors in Java, and how to mitigate the
issue.

Thanks to Eric McCarthy for trying ATJ on an example that exhibited a stack
overflow error (due to a non-tail-recursive ACL2 function being called on a
large input) and for subsequent discussion that pointed at the need to explain
stack overflows in the user documentation.