[Java] Add native implementation of NONNEGATIVE-INTEGER-QUOTIENT.
Now AIJ includes a native Java implementation of this function, and ATJ ignores
this function's ACL2 definition and instead generates Java code that uses the
AIJ native implementation.

NONNEGATIVE-INTEGER-QUOTIENT is a built-in ACL2 function with raw Lisp code.
Before this change, ATJ was treating it like other functions, based on its
recursive ACL2 definition. However, it was observed that the generated Java code
runs out of stack space when the method to which this function was translated
was called with large values -- the calls were through MOD, whose ACL2
definition calls FLOOR, whose ACL2 definition calls
NONNEGATIVE-INTEGER-QUOTIENT. Now that this function is natively and efficiently
implemented (in terms of Java BigInteger's division), the stack overflow no
longer occurs.

Thanks to Eric McCarthy for reporting the issue when using ATJ on Keccak.