[Std/util] Add a DEFISO macro tool.
This macro serves to verify and record an isomorphic mapping between two domains
(i.e. predicates). The user supplied the two domains and the two purported
isomorphisms (i.e. functions), and the macro generates applicability conditions
(i.e. theorems to prove) saying that the isomorphisms map the first domain to
the second and vice versa and are mutual inverses on the domains. There are also
optional applicability conditions related to guards. If all the applicability
conditions prove, the tool automatically proves injectivity theorems of the two
isomorphisms over the two domains, and records all the theorems, along with
domains and isomorphisms, in a table.

Although this macro was developed to support APT's isomorphic data
transformation (which will be open-sourced soon), it is more general. This is
why I'm putting it under (Kestrel's extensions of) Std/util.

The tool comes with user documentation, design notes with meta proofs,
illustrative template proofs, and extensive tests (many of which
template-based).