[Java] Add a native Java implementation of STRING-APPEND.
In ACL2, the :EXEC part of STRING-APPEND calls CONCATENATE. This macro runs via
raw Lisp code, but its ACL2 definition is in terms of STRING-APPEND. Thus, when
using ATJ to generate Java code with :GUARDS T (i.e. assuming that guards are
satisfied, which in particular means that the :EXEC parts of MBEs are translated
to Java), the Java code was unexpectedly looping forever.

By extending AIJ with a native Java implementation of STRING-APPEND, and
adapting ATJ to rely on that without generating Java code for STRING-APPEND, the
circularity is avoided. Furthermore, the native implementation is faster also
when executing Java code generated with :GUARDS NIL (i.e. executing in the
logic, which in particular means that the :LOGIC parts of MBEs are translated to
Java), which has no infinite loops in this case. The only drawback is
potentially slightly less assurance, but the native Java implementations of this
and other ACL2 funtions (including the primitive ones) would have to be formally
verified eventually, for very high assurance.