[Java] Start using the pseudo-term fixtype in ATJ.
Most of the ATJ is in logic mode and guard verified, but some guard and return
proofs are slow due to the fact that the standard pseudo-term operations are
designed for fast program-mode execution. Using the pseudo-term fixtype
operations instead in the slowest input processing function in ATJ, the file
takes about 1/2 of the time to certify.

It is planned to use the pseudo-term fixtype operations in other ATJ code as
well, to speed up the guard and return proofs.