Add new U command in an extension of acl2-doc.  Fix license headers in emacs/ files.
Quoting :doc note-8-3:

  The [ACL2-doc] browser for ACL2+books documentation can be extended
  with a new command, U, to open a URL (or in some cases, a file) in
  a browser.  See file emacs/acl2-doc-open-url.el for more
  information.

The fix for license headers in emacs/ files is to clarify that LICENSE is
to be found in the main ACL2 source directory.

NOTE: I might incorporate the new U command into acl2-doc proper.  In
the meantime, I'd certainly be interested in feedback -- feel free to
email questions to me!