[APT] Improve ISODATA.
Have ISODATA also generated a (local, for now) theorem that expresses the new
function in terms of the old function. The proof is generated according to the
template proofs in the design notes.

This theorem will be used to simplify the generated proof of the (exported)
theorem that expresses the old function in terms of the new function.