[Java] Use more pseudo-term FTY operations in ATJ.
Using the pseudo-term FTY operations in the ATJ mutually recursive functions to
type-annotate terms, instead of the built-in term operations, reduces the time
to process these ATJ functions by about an order of magnitude -- from about 90
seconds to about 9 seconds. The processing includes termination, return type
theorems, and guard verification.

This cuts the time to certify the ATJ pre-translation file roughly in half.

More optimizations like this are planned.