[Java] Add an abstract model of floating-point literals.
This is in line with the fact that we have an abstract model of floating-point
values, instead of modeling the absence of floating-point values and literals as
an approximation, which may be difficult to lift in the future if many
definitions and proofs were to depend on that absence. In this modeling approach
we say that floating-point values and literals are there, but we just don't know
much about them (yet).