[APT] Improve ISODATA.
Generate a wrapping IF test for the transformed arguments being in the new
representation also in the non-recursive function case, so that this case is
uniform with the recursive function case. This also provides a better context
for subsequent transformations, such as simplification.

Adapted the (local, for now) new-to-old theorem accordingly.

Adapted a few tests as needed.

Updated user documentation to include the new wrapping test.

Also add a :COMPATIBILITY option that, when T, generates non-recursive functions
without the wrapping test. This will facilitate the adaptation of existing
developments that may rely, perhaps implicitly, on the absence of the wrapping
test.