[Std/system] Improve CHECK-MV-LET-CALL.
Extend to handle cases in which the lambda expressions in the translated MV-LET
have additional trivial variables (i.e. applied to themselves) to keep the
lambda expressions closed. In cases like this the variable bound by the outer
lambda to the multi-value may not be the symbol MV, if that occurs free in the
untranslated term. So the utility now also returns the symbol used for that
variable (often MV, but not always).

Adapted the existing tests to the new result, and added new tests that exercise
the situations with extra variables described above.