[Java] Improve ATJ.
Replace calls of ATJ-CHECK-MV-LET-CALL with the recently improved
CHECK-MV-LET-CALL and with a newly added FTY-CHECK-MV-LET-CALL that is a simple
wrapper of CHECK-MV-LET-CALL that fixes the argument. Eliminate
ATJ-CHECK-MV-LET-CALL.

This not only improves the modularity of the implementation, but also takes
advantage of the recent improvements to CHECK-MV-LET-CALL that handle less
common but possible situations with translated MV-LETs.