[Std/strings] Move tests to separate files.
As discussed on the ACL2 Books mailing list.

In one instance, namely the file pretty.lisp, the assertions looked more
validations that the implementation relies on, rather than actual tests. So I
left them there, but replaced ASSERT! from Std/testing with the built-in
ASSERT-EVENT, thus avoiding the dependency on Std/testing anyhow.