[Java] Improve the representation of the ATJ types.
Instead of using just keywords to represent internally the ATJ types, use
structured (fix)types.

The keywords are retained as the user-facing representation in the
DEF-ATJ-MAIN/OTHER-FUNCTION-TYPE macro.

This change of representation required numerous changes in the code,
unsurprisingly. In the process, some aspects of the code were slightly improved.