[APT] Extend ISODATA's input processing.
Now the input processing code allows the specification of results of multi-value
functions to be isomorphically transformed. This is done via keywords :RESULT1
throught :RESULTm, where m is the number of results of the target function.  If
the function is single-valued, :RESULT is allowed as a synonym of :RESULT1.

This is a first step towards supporting the isomorphic transformation of results
of multi-value functions, but the rest of the implementation remains to be
extended. Thus, the input processing code contains a transitional ending step in
which we ensure that the function is single-valued if any (in fact, the only)
result is being transformed, for compatibility with the rest of the
transformation.

The rest of the implementation will be extended soon to handle the
transformation of results of multi-value functions. When that is done, the
transitional ending input processing step mentioned above will be removed, and
the user documentation will be extended to describe the newly added feature.