[event macros] Simplify an input processing utility.
This is for the :HINTS input of event macros.

This commit replaces it with the newer, simpler version with the final $ in its
name, which is now just a synonym to keep the existing calls (inside and outside
the community books) working. The $ synonym will be removed once those calls are
changed to call the non-$ function.