[Java] Extend ATJ's translation.
Treat the deconstructors of Java primitive types specially. Generate Java code
that converts them to corresponding Acl2Symbol or Acl2Integer values.

These are explicit conversions, in the sense that these deconstructors are
present in the original ACL2 code, prior to it being type-annotated (which is
when implicit conversions are inserted).