[Java] Robustify some ATJ type annotations.
Rewrap the arguments of a function call also when function types are found for
the minimum input types greater than or equal to the argument inferred types.
Not rewrapping is adequate when up-conversions between ACL2 types are no-ops in
Java as they are now, but will be needed when those up-conversions are no longer
no-ops, in future extensions in which there may be more flexible mappings
between ATJ types and Java types.