[Java] Fix bug in ATJ.
This was in the variable reuse pre-translation step: certain variables that
could have been reused were not reused. This bug did not affect soundness, as it
is always safe to not reuse variables. However, it was causing variable reuse
opportunities to be missed.

This came up in a test involving array updates, whose reusability as variables
is ensured by the array analysis pre-translation step. Instead, one of the tests
showed arrays variables being unncessarily renamed (i.e. not reused).