[Java] Slightly extend ATJ's model of Java primitive arrays.
Add predicates to check that a Java int is in the range of a Java primitive
array -- 8 predicates, one for each primitive type.

These predicates are useful in guards of functions that may be targets of
transformations aimed at generating Java code that manipulates Java primitive
arrays.

Some code in the array model and in some tests has been made shorter by using
the new predicates.