[Java] Finish formalizing Java literals.
Add a (tagged sum) fixtype for all the Java literals, which puts together all
the separately formalized types of literals. Except for the floating-point
literals which are currenly modeled via an abstract placeholder, the model is
accurate in the sense that it preserves all the information from the concrete
syntax (but this should be formally proved eventually), albeit in a more
conveniently "packaged" form.