[Java] Remove ATJ's representation check for Java primitive values.
Because of the new way in which the conversions between Java primitive values
and ACL2 built-in values work, the internal representation of our formalization
of the Java primitive values is no longer "exposed". In other words, ATJ's code
generation functions treat Java primitive values (and arrays) abstractly: that
is, the representation of the types could be changed without changing any other
part of the ATJ code.