[FTY] Add a new macro DEFBYTE-IHS-THEOREMS.
This macro takes as input the name of a fixtype introduced by DEFBYTE, and
generates theorems (currently just one) that relate the fixtype to functions in
IHS. The macro can be extended to generate more theorems as needed.

The reason for having a separate macro, instead of having DEFBYTE directly
generate these theorems, is to avoid necessarily including (some of) IHS when
DEFBYTE is used.

Perhaps at some point this DEFBYTE-IHS-THEOREMS macro could be superseded by an
enhanced DEFBYTE that has "pluggable" features along the lines of macros like
DEFLIST, which generates more theorems when more books are included.