[APT] Extend ISODATA.
Now this transformation generates non-locally also the theorem saying
that the new function maps arguments in the new representation to
results in the new representation. (This happens only if some result
is transformed.)

New inputs :NEWP-OF-NEW-NAME and :NEWP-OF-NEW-ENABLE have been added
to control the name and enablement of the theorem.