[Java] Extend ATJ's post-translation.
In particular, extend the array write method call remove step to check that
there are no more such calls left, as expected (but not formally proved
currently). This will let us avoid the generation of such methods, only
ephemerally generating calls in the translation step, which are then eliminated
in post-translation.