[Java] Extend AIJ.
Add variants of some natively implemented ACL2 functions that return Java
booleans instead of ACL2 booleans.

This is in support of a planned extension of ATJ that will allow ACL2 booleans
to be mapped directly to Java booleans, instead of to (Java representations of)
the ACL2 symbols T and NIL. That is the extension will support the generation of
"unwrapped booleans".

The newly introduced method variants are for ACL2 functions that are natively
implemented by AIJ and that return (ACL2) booleans, such as the methods that
implement CONSP and EQUAL. The newly added methods return boolean instead of
Acl2Symbol. These are not overloaded methods, because only the result type
differs, so they must have different names.