[Java] Improve the ATJ type mechanism.
"Move" the meet semilattice from the Java types to the ATJ types. This was
motivated by an upcoming extension in which ACL2 booleans are mapped to Java
booleans: while (the upcoming) :ABOOLEAN is a subtype of :ASYMBOL, it is not the
case that the Java type boolean is a subtype of the Java type Acl2Symbol. So, to
make things work with this extension, it was necessary to calculate greatest
lower bounds on ATJ types rather than Java types.

The resulting structure is cleaner and more satisfactory, as an added benefit.