[Java] Improve the code generated by ATJ.
Now, when :DEEP is NIL and :GUARDS is T, ACL2 booleans are automatically mapped
to Java booleans when possible.

In order to accomplish this, the ATJ type system has been extended with a new
type :ABOOLEAN for ACL2 booleans. This is below :ASYMBOL in the partial order,
and consits of the ACL2 symbols T and NIL, i.e. the ACL2 values of type
BOOLEANP. ATJ's ACL2-to-Java translator maps :ABOOLEAN to the Java boolean type,
automatically inserting conversions between Java booleans and Acl2Symbol objects
(to which the supertype :ASYMBOL is mapped) as needed, as inferred by ATJ's type
analysis. With this extension, it is no longer the case that the subtype
relation on ATJ types always implies the Java subtype relation on the
corresponding Java types.

This feature is distinct from the ability to explicitly target Java primitive
types from suitably phrased ACL2 code. The latter ability is based on the
:JBOOLEAN type, and must be explicitly arranged for in the ACL2 code, as done
for the other Java primitive constructs. In contrast, the mapping of :ABOOLEAN
to Java booleans happens automatically.

This feature should make the generated Java code more efficient and idiomatic in
general.