[Java] Formalize all the Java types.
As syntactic entities.

This definition of the fixtype JTYPE with recognizer JTYPEP is not yet included
in [books]/kestrel/java/language/syntax.lisp, and thus in
[books]/kestrel/java/language/top.lisp, and thus in
[books]/kestrel/java/top.lisp, because, in the latter, it would conflict with
ATJ's own definition of JTYPEP. The plan is to have ATJ soon use the new
(better, more complete) definition of JTYPEP just added, removing the other one,
which was temporary and more approximate (but adequate to ATJ's needs). Instead
of renaming one or the other JTYPEP, we temporarily exclude the new one from the
manual (via the syntax.lisp and top.lisp files).