[Java] Improve code generated by ATJ.
Add a post-translation (i.e. Java-to-Java) step to simplify Java expressions of
the form

  true ? <then> : <else>
  false ? <then> : <else>

to just <then> and <else> respectively. These expressions are generated, in
particular, by the ACL2-to-Java translation step when quoted T and NIL are used
as symbols and not booleans (e.g. they are passed as arguments to a function
whose ATJ input type is :ASYMBOL), so that Java code is inserted to convert the
Java booleans to the Java Acl2Symbol objects for T and NIL.

Thanks to Limei Gilham for providing an example of these generated expressions.