Fixed read-string implementation for SBCL 2.0.5.
Quoting :doc note-8-4-books:

  Standard IO Library (see [STD/IO])

  Modified the raw-Lisp implementation of [read-string] to catch errors
  successfully, as intended.  This could fail in SBCL 2.0.5
  (implementation note: because the error message referenced a string
  stream that was already closed).

These two tests were failing in books/std/io/read-string-tests.lisp:

(test-fail "#\wtf ")
(test-fail "#\Return ")

The problem was that the handler-case form in function read-string-fn
was reporting an error that mentioned a string-stream that was outside
the scope of its binding in with-input-from-string.  (Thanks to Stas
Boukarev on the sbcl-help list for pointing this out.)  This isn't a
problem for the other Lisps that host ACL2, because they don't mention
the stream in their error.