[Java] Add some native implementation variants to AIJ.
These are native Java implementations of ACL2 functions that take or return
characters. They are variants that take or return Java characters instead of
ACL2 characters.

Their purpose is to support an upcoming extension of ATJ in which ACL2 characters
will be mapped to Java characters (below 256), automatically converting them to
Acl2Character objects as needed. This is similar to a recent extension of ATJ to
map ACL2 booleans to Java booleans.