[ACL2PL] Start library formalizing the ACL2 programming language.
This contains an initial formalization of an evaluation semantics of ACL2. As
mentioned in the top-level XDOC topic of this library, this is just the author's
work, not an official semantics; for an official semantics, see :DOC EVALUATION.

This ACL2 evaluation semantics is formalized as a small-step operational
semantics, with an explicit stack of frames that may grow or shrink at each
step. The small-step approach is more flexible, in particular allowing us to
describe properties of non-terminating computations; a big-step semantics can be
derived from the small-step one. The evaluation is currently formalized at the
level of the translated terms that are the unnormalized bodies of the functions.
Evaluation is "in the logic" for now, i.e. ignoring guards completely. Side
effects are not modeled for now. Other than that, the model captures all ACL2
primitive and defined functions. The execution of the ACL2 primitive functions,
which have no bodies, is formalized as single evaluation steps.

Future versions of this formalizated semantics may include guards and their
various treatments (based on guard checking mode), side effects, and evaluation
at the level of untranslated terms, including macros.

The motivation for formalizing an evaluation semantics of ACL2 is to generate
proofs of correctnes of the Java code generated by ATJ: the proofs will involve
this semantics as well as a Java semantics that is also being developed.
However, a formalization of an ACL2 evaluation semantics in the ACL2 logic is
more general and has other uses. In particular, it assigns formal meaning to
program-mode ACL2 functions (including ones that do not terminate, intentionally
so, as used, for instance, in server or control loops), making it possible to
prove properties of program-mode functions in the ACL2 logic (at the meta
level), just like one can prove properies of programs written in other
programming languages.