[Java] Extend ATJ.
In the shallow embedding with guards, now ACL2 strings are mapped to Java
strings (i.e. java.lang.String objects) when possible, automatically converting
between Java strings and Acl2String objects when needed.

This is similar to the capability, added previously, to map ACL2 booleans and
characters to Java booleans and characters.

This happens automatically. It is a distinct capability from the one to
explicitly target Java types and operations via their ACL2 models.