[Java] Add and use a native Java implementation of CHAR.
Extend AIJ with a native implementation of the ACL2 built-in function CHAR. This
function's unnormalized body converts the string to a list of characters and
then calls NTH; but in execution it runs raw Lisp code. This native
implementation accesses the underlying Java string directly, which should make
execution faster in the Java code. This native implementation consists of a few
variants, for different argument and result types, similarly to other native
implementation of ACL2 built-in functions.

Extend ATJ to support the mapping of CHAR to its native implementation.

Add tests for (the native implementation of) CHAR.