[Java] Fix bug in ATJ.
After adding a pre-translation step that restores ANDs from IFs of suitable
form, the existing variable reuse pre-translation step had become incorrect (in
certain situations) because it was considering AND as a strict function (because
it was only considering IF a non-strict function). One possible fix was to
modify the variable re-use pre-translation step to regard AND as non-strict. But
a simpler fix is to move the AND reconstruction pre-translation step after the
variable reuse step: this is the fix chosen here.

In fact, the AND reconstruction step has been moved also after the variable
renaming pre-translation step, which is also affected by the non-strictness of
AND.

Thanks to Stephen Westfold for producing an ACL2 function that exhibited the
issue when translated to Java, and to Eric Smith for noticing the problem when
trying to compile the generated Java file (which resulted in a compiler error).