[Java] Fix bug in worklist algorithm.
When descending in either argument of MBEs to collect the functions to either
generate Java code for or check for absence of side effects, the algorithm
neglected to also descend into the other argument for checking absence of side
effects (as correctly done for PROG2$ and PROGN). Now the algorithm correctly
descends into the other argument of MBE, in analogy to the treatment of PROG2$
and PROGN.

This only affects user input validation, not the generated Java code.