[Java] Improve the code generated by ATJ.
In the shallow embedding approach, avoid generating Java classes with just
constant fields for imported symbols.

In general, the Java class for an ACL2 package includes static final fields for
(1) the symbols in that pagkage that are referenced by at least one ACL2
function that we translate to Java and (2) the symbols imported by that package
that are referenced by at least one ACL2 function that we translate to Java.

However, if a package has no function that we translate to Java, i.e. if the
class has no methods, then there is no need for the fields (2) described above:
we just need the fields (1). This is because the fields for the imported symbols
are only useful to be referenced by code in the class. This commit thus avoids
the fields (2) when the class has no methods.

This means that, for all of our current tests and applications, we avoid the
generation of several classes for ACL2 packages that are in the ACL2 world when
ATJ is invoked, but that contain no functions that we translate to Java, and no
symbols referenced by those functions.