[Java] Improve ATJ's tail recursion elimination.
Now this post-translation step handles better the case in which the tail
recursive call is in the rightmost conjunct of a right-associated && nest.
Previously, the code was correct but odd-looking, with a recursive call inside a
loop. Now we generate a loop without recursive calls. The && nest is suitably
broken up into an if-else.