[APT] Simplify the TAILREC transformation.
Remove the :NON-EXECUTABLE option, since there is no evidence that this option
is necessary or even useful. Now the new function is non-executable iff the
target function is; the wrapper, if generated, is always executable.

If we ever find it useful to change the executable status of a function, a
specific transformation could do that, which seems a better factoring than
having each single transformation provide an option to do that.