[Java] Add an option to ATJ.
A new option :IGNORE-WHITELIST has been added, NIL by default. When T, ATJ will
ignore the whitelist of ACL2 functions with raw Lisp code, and generate Java
code for any of them (so long as they have unnormalized bodies). This means that
any side effects in the raw Lisp code won't be reproduced in Java, but this may
be fine if, for instance, the functions with raw Lisp code are unreachable under
the guards. In any case, the user has to explicitly set this option.