[APT] Extend the TAILREC transformation.
Have it also generate a theorem that rewrites the new function in terms of the
old function, along with a theory invariant preventing this from being enabled
at the same time as the theorem that rewrites the old function in terms of the
new function (and also some other invariants related to the theorems about the
wrapper function, when this function is generated).

This new-to-old theorem was already being generated internally (i.e. locally to
the ENCAPSULATE), and in fact is the key lemma for the correctness of the new
function, but now it is also exported.

New inputs :NEW-TO-OLD-NAME and :NEW-TO-OLD-ENABLE have been added to control
the name and enablement of this theorem.