[APT] Improve the TAILREC transformation.
Simplify the :OLD-TO-NEW-NAME input to be either a symbol to use for the
theorem, or a keyword used to separate the names of the old and new function in
the theorem name (e.g. :-TO-). Furthermore, if this input is absent, it is the
keyword in the APT defaults table.