[APT] Extend the TAILREC transformation.
Add a new variant, :ASSOC-ALT, which is an alternative to the existing :ASSOC
variant. The extension involves the design notes, the user documentation, the
implementation, and the tests/examples. In particular, now the transformation
can successfully transform the function that returns the minimum of a non-empty
list of rational numbers (and of course all the functions with a similar
pattern). The binary operator here is ACL2's MIN, which is associative but has
no identity elements. This alternative associative variant relaxes some
applicability conditions but strengthens the associativity one to be
unconditional, which is true for MIN.