[APT] Simplify and improve an XDOC constructor.
This is the one for the :WRAPPER-NAME input of transformations.

We simplify it by removing the macro parameter that controls whether the wrapper
is optional or mandatory: the rationale is that for now it seems that we'll
always want the wrapper to be optional; if this turns out to be the case, we
will introduce another XDOC constructor, or parameterize this one again.

We improve it by referencing the XDOC topic about function name generation in
APT.