[APT] Simplify the RESTRICT transformation.
Remove the :NON-EXECUTABLE option. There is no evidence that this option is
useful so far. If the non-executable state of a function needs to be changed in
the course of an APT derivation, it may be better to have a simple dedicated
transformation than to have every transformation support a :NON-EXECUTABLE
option.