[ecurve] Change the representation of points.
The old representation, (x . y), which used (0 . 0) for the point at infinity,
was not adequate for those curves in which (0, 0) happens to satisfy the curve
equation, but is not the point at infinity.

The new representation is the union of (x . y) and :infinity, i.e. where the
point at infinity is represented explicitly.

A number of definitions, theorems, and proofs had to be adapted to this change,
including in the specification of ECDSA signature.

The representation of the secp256k1 curve points has not been changed. It may be
changed in the future, to make it more uniform with the new general
representation of points. A file with explicit conversions between the general
point representation and the secp256k1 point representation has been also added.