[APT] Improve an input processor.
This is for the :NEW-NAME input of APT transformations that do not generate a
wrapper function (i.e. that do not have also a :WRAPPER-NAME input). The input
processor has been improved to be more similar to the joint one for the
:NEW-NAME and :WRAPPER-NAME inputs: the new version takes a list of names to
avoid as additional input, and returns an updated list as additional output;
furthermore, the new version uses a more accurate name freshness check.