Updated :doc books-certification to reflect the new meaning of "make all" for book regressions.  Fixed a bit of obsolete wording in ./GNUmakefile.
I don't claim to have caught all places where the documentation needs
updating to reflect the recent changes in books/GNUmakefile, but I
happened to notice one in :doc books-certification, which I've fixed
as mentioned above.

I hope that others will keep an eye out for additional such
documentation updates that need to be made due to yesterday's
books/GNUmakefile changes.

Thanks to Eric Smith for pointing out the obsolete wording in
a comment in ./GNUmakefile.