Print certification failure tails atomically
When running a regression, books print only two lines when they are certified
successfully (the "Making ~" and the "Built ~" lines), but when they fail, they
print the last 300 lines of their `.cert.out` file to aid in debugging.  Before
and after this 300-line tail are header and footer lines identifying which book
the failure log is coming from.

When running regressions with a large `-j` number (i.e. many builds running in
parallel), sometimes output starts interleaving a bit, and the headers and
footers can get separated from the 300-line tail, making it harder to tell at a
glance which file a 300-line tail is coming from.

While this is sort of inevitable (short of taking full control over
stdout/stderr of the build processes and doing some smart buffering), one
low-hanging fruit is the fact that the header and footer are printed in
`make_cert_help.pl` while the 300-line tail is printed in two new chained
processes spawned between those two `print` statements, which has to do disk I/O
to seek to the last 300 lines of the appropriate `.cert.out` file, adding some
latency.

This commit makes `make_cert_help.pl` create and format the 300-line tail in
memory before dumping the header, the tail, and the footer all in one `print`
statement.  Hopefully this will mitigate the interleaving to some extent...