[event macros] Add a proof preparation utility.
This can be used by event macros to change certain ACL2 settings that may
"sabotage" generated proofs that are designed to always work.

Currently the utility disables default hints and override hints, and disables
the special treatment of MV-NTH being expanded in proof even if disabled (the
disabling of the special treatment is achieved via a system attachment recently
added to ACL2).