[JSON] Start a library for JSON.
Currently this library just contains fixtypes for an abstract syntax of JSON.

This abstract syntax should be isomorphic to the one generated by a JSON parser
written by Eric Smith, which is currently in the Kestrel internal files but may
be open-sourced soon. However, that abstract syntax does not use fixtypes. So
the one defined here is useful when writing "fixtyped" code.