[APT] Add a divide-and-conquer list-fold transformation.
This is a start towards a class of transformations to apply algorithm schemas to
non-executable specifications. Given a functional specification consisting of an
input/output relation (e.g. consisting of a precondition and postcondition),
this transformation partially determines the solution as a recursive function
that follows a divide-and-conquer schema (i.e. decompose the problem, solve the
sub-problems, and compose the sub-solutions to obtain a solution).

Currently the algorithm schema is limited to list fold, i.e. the decomposition
is CDR (i.e. the same recursive structure of many list operations, such as LEN).
Support for additional divide-and-conquer schemas is planned.

The partial determination of the solution via the schema amounts to decomposing
the initial specification into sub-specifications that can be recursively
subjected to APT transformations.

This divide-and-conquer transformation is currently realized using SOFT to
represent specifications and algorithm schemas, with specifications as 2nd-order
functions and refinement as backward implication, as described in the ACL2-2015
Workshop paper on SOFT. In the future, this transformation may be extended to
represent specifications and algorithm schemas via the built-in APPLY$.

This whole approach should work for other algorithm schemas besides
divide-and-conquer.

This commit includes not only the implementation, but also design notes, user
documentation, tests (some of which schematic), and explicated proof templates.