[Java] Improve AIJ Makefile.
Use files instead of directories as dependencies of the targets to generate
Javadoc HTML.

On macOS (the latest), using directories worked fine in the sense of avoiding
rebuilding them when up to date. However, this did not seem to work on (at least
some instances of) Linux. Using files as targets (specifically, the index.html
files in the directories, seems to work now).

This saves a bit of build time, and reduces unnecessary screen output.