[APT] Extend the SOLVE transformation.
Add support for the ACL2 rewriter as the solving method.

Thanks to Matt Kaufmann for creating REWRITE$, a user-friendly programmatic
interface to the ACL2 rewriter. The SOLVE transformation uses REWRITE$ to invoke
the ACL2 rewriter.

The method is available in a separate file that depends on REWRITE$, so that
including just SOLVE does not automatically force the inclusion of
REWRITE$. This is farily minor for REWRITE$, but it is part of the modular
architecture of SOLVE, which is useful for other solving methods that bring more
dependencies, such as the Axe rewriter (already supported, but currently only
available inside Kestrel), and SMT and SAT solvers (in the future).