[FTY] Add the :EXECUTABLEP option to DEFFIXTYPE.
This was discussed with Sol Swords.

The documentation of DEFFIXTYPE mentions :EXECUTABLEP, but the implementation
uses :EXECP. The former seems preferable to the latter, so we want to replace
the latter with the former in the implementation.

As suggested by Sol, we do that in three steps:

1. We add :EXECUTABLEP without removing :EXECP; the default T, its conjoining
with EXECP, and the disjoining of its negation with (NOT EXECP), should ensure
no change of behavior in existing calls. The full regression passes.

2. We switch existing calls to use :EXECUTABLEP, whose treatment in the
implementation is symmetric to :EXECP, so again there should be no change in
behavior.

3. We remove :EXECP.

This commit does step #1.