Improved proof-builder message that could occur when backchaining to a false FORCE or CASE-SPLIT hypothesis.
Thanks to Mihir Mehta for feedback leading to this improvement.

(I added a comment in (defxdoc note-8-4 ...) but not a release note,
since this is relatively minor.)

Here's an example.

(defstub f (x) t)
(defstub g (x) t)
(defaxiom f-holds (f x))
(defaxiom f->g (implies (case-split (not (f x))) (g x)))
(verify (g a))
(s)

;;; Old:
*** NO CHANGE *** -- A false assumption was encountered from applying
the rune (:REWRITE F->G) to the target (G A).

;;; New:
*** NO CHANGE *** -- When applying the rune (:REWRITE F->G) to the
target (G A), a hypothesis of the form (FORCE ...) or (CASE-SPLIT ...)
was later found to be false.