Complete the implementation per suggestions by @MattKaufmann & @kini
A note for Matt and Keshav: thanks for your input! The new utility is
now called find-defs, and both find-lemmas and find-defs are now
variants of the new find-events. It's OK for either of you to merge
the pull request now - hopefully the tests I have added will assure
that the existing functionality still works as expected.