Added to books/doc/cert.acl2 to support :doc note-8-4-books at the terminal.  Added open-trace-file!.  Tweaked defwarrant for apply$ primitives.  Cleaned up a bit of source code.  Added a dependency.  Tweaked an error message.  Deleted a comment.
Quoting :doc note-8-4-books:

  See [open-trace-file!] for a variant of [open-trace-file] suitable for
  use within [make-event].

Thanks to Shilpi Goel for requesting open-trace-file!.  Note that the
symbol trace-co was added to the list value of ACL2 source constant
*protected-system-state-globals* in support of that change, so that
the trace file stays open outside the make-event call.

Quoting :doc note-8-4:

  Improved [defwarrant] to be a no-op for [apply$] primitives.

Quoting a comment in :doc note-8-4:

; Removed an unnecessary argument (specifically, pt) from ACL2 source function
; rewrite-clause-type-alist, since the only call of rewrite-clause-type-alist
; s by rewrite-clause and pt was the same as the :pt field in the local-rcnst
; passed there to rewrite-clause-type-alist,.

In books/, added dependency of projects/apply/base.lisp on
build/first-order-like-terms-and-out-arities.certdep -- I'm reasonably
confident (but not 100% sure) that it's necessary.

; Tweaked an error message when the :check-expansion or :expansion? argument of
; ~ make-event would normally be a consp (source function
; chk-embedded-event-form).  Thanks to Mihir Mehta for feedback leading to this
; change.

Deleted a comment in forward-chain-top about not rewriting with a
type-alist, since the paragraph above it makes it clear that we have
parent trees around (which should avoid tail-biting, which is
presumably the concern in the paragraph below it).

Passed the usual "everything" regression and also (before the last two
trivial mods) "make devel-check".  Note: My (new) laptop has glucose
installed but not abc, ipasir, or smtlink.  I might install smtlink in
the future.